Fix warnings by MSVC 64 in qtranslator.cpp.
kernel\qtranslator.cpp(880) : warning C4267: 'initializing' : conversion from 'size_t' to 'const uint', possible loss of data
kernel\qtranslator.cpp(881) : warning C4267: 'initializing' : conversion from 'size_t' to 'const uint', possible loss of data
kernel\qtranslator.cpp(882) : warning C4267: 'initializing' : conversion from 'size_t' to 'const uint', possible loss of data
kernel\qtranslator.cpp(975) : warning C4267: 'initializing' : conversion from 'size_t' to 'const int', possible loss of data

Task-number: QTBUG-39757

+++ src/corelib/kernel/qtranslator.cpp
@@ -877,9 +877,9 @@ static QString getMessage(const uchar *m, const uchar *end, const char *context,
 {
     const uchar *tn = 0;
     uint tn_length = 0;
-    const uint sourceTextLen = strlen(sourceText);
-    const uint contextLen = strlen(context);
-    const uint commentLen = strlen(comment);
+    const uint sourceTextLen = uint(strlen(sourceText));
+    const uint contextLen = uint(strlen(context));
+    const uint commentLen = uint(strlen(comment));
 
     for (;;) {
         uchar tag = 0;
@@ -972,7 +972,7 @@ QString QTranslatorPrivate::do_translate(const char *context, const char *source
             return QString();
         c = contextArray + (2 + (hTableSize << 1) + (off << 1));
 
-        const int contextLen = strlen(context);
+        const uint contextLen = uint(strlen(context));
         for (;;) {
             quint8 len = read8(c++);
             if (len == 0)
